requirements.txt
Flask>=2.2.2
Werkzeug>=2.2.2
python-dotenv==1.0.0
elasticsearch==8.17.0
sentence-transformers==2.2.2
huggingface-hub==0.11.1
openai==0.27.0
psycopg2-binary==2.9.6
numpy==1.23.5
redis